Refactor HttpClient integration tests to ConsoleMF
